requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
httpx==0.27.2
python-multipart==0.0.9
Jinja2==3.1.4
PyJWT==2.9.0
pyotp==2.9.0
itsdangerous==2.2.0
qrcode==7.4.2
